使用time命令统计程序执行时间与内容

+++ src/main/java/com/code/codesandbox/pojo/CodeResult.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CodeResult {
     private String result;
-    private Long runtime;
+    private Double runtime;
     private Double memory;
 }
